+++ src/aliceVision/calibration/distortionEstimation.hpp
@@ -9,48 +9,11 @@
 #include <aliceVision/numeric/numeric.hpp>
 #include <aliceVision/camera/Pinhole.hpp>
 #include <aliceVision/camera/Undistortion.hpp>
-
 #include <vector>
 #include <memory>
 
 namespace aliceVision {
 namespace calibration {
-
-struct PointWithScale
-{
-    Vec2 center;
-    double scale;
-};
-
-/**
- * @brief Set of 2D points that belong to the same line.
- *
- * The 2D points correspond to real-world points that are aligned and evenly spaced
- * observed with a camera which applies some distortion on them.
- *
- * Therefore these 2D points may not actually be aligned and evenly spaced,
- * and this difference with the ideal line model is used to estimate distortion.
- */
-struct LineWithPoints
-{
-    double angle;
-    double dist;
-    std::vector<PointWithScale> points;
-};
-
-
-/**
- * @brief a pair of points coordinates
- * 
- * One vector for the distorted coordinates
- * One vector from the undistorted coordinates
-*/
-struct PointPair
-{
-    Vec2 distortedPoint;
-    Vec2 undistortedPoint;
-    double scale;
-};
 
 /**
  * @brief Statistics on distortion parameters estimation error.
@@ -63,66 +26,6 @@ struct Statistics
     double lastDecile;
     double max;
 };
-
-/**
- * @brief Estimate the undistortion parameters of a camera using a set of line aligned points.
- *
- * This algorithms minimizes a distance between points and lines using distortion.
- *
- * @param[out] undistortionToEstimate Undistortion object with the parameters to estimate.
- * @param[out] statistics Statistics on the estimation error.
- * @param[in] lines Set of line aligned points used to estimate distortion.
- * @param[in] lockCenter Lock the distortion offset during optimization.
- * @param[in] lockDistortions Distortion parameters to lock during optimization.
- * @return False if the estimation failed, otherwise true.
- */
-bool estimate(std::shared_ptr<camera::Undistortion> undistortionToEstimate,
-              Statistics& statistics,
-              std::vector<LineWithPoints>& lines,
-              const bool lockCenter,
-              const bool lockAngles,
-              const std::vector<bool>& lockDistortions);
-
-
-/**
- * @brief Estimate the undistortion parameters of a camera using a set of pair of undistorted/distorted points.
- *
- * This algorithms minimizes a distance between points and points using distortion.
- *
- * @param[out] undistortionToEstimate Undistortion object with the parameters to estimate.
- * @param[out] statistics Statistics on the estimation error.
- * @param[in] pointpairs Set of pair of points used to estimate distortion.
- * @param[in] lockCenter Lock the distortion offset during optimization.
- * @param[in] lockDistortions Distortion parameters to lock during optimization.
- * @return False if the estimation failed, otherwise true.
- */
-bool estimate(std::shared_ptr<camera::Undistortion> undistortionToEstimate,
-              Statistics& statistics,
-              const std::vector<PointPair>& pointpairs,
-              const bool lockCenter,
-              const std::vector<bool>& lockDistortions);
-
-/**
- * @brief Estimate the undistortion parameters of a camera using a set of pair of undistorted/distorted points.
- *
- * This algorithms minimizes a distance between points and points using distortion.
- *
- * @param[out] undistortionToEstimate Undistortion object with the parameters to estimate.
- * @param[out] statistics Statistics on the estimation error.
- * @param[in] pointpairs Set of pair of points used to estimate distortion.
- * @param[in] lockCenter Lock the distortion offset during optimization.
- * @param[in] lockDistortions Distortion parameters to lock during optimization.
- * @param[in] T the in/out transformation matrix
- * @return False if the estimation failed, otherwise true.
- */
-bool estimate(std::shared_ptr<camera::Undistortion> undistortionToEstimate,
-              Statistics& statistics,
-              const std::vector<PointPair>& pointpairs,
-              const bool lockCenter,
-              const std::vector<bool>& lockDistortions,
-              Eigen::Matrix3d & R, 
-              Eigen::Vector3d & t,
-              double &scale);
 
 }  // namespace calibration
 }  // namespace aliceVision
